Avoid using cached ColorSpace after moving window to other display
