+++ os/win/color_space.h
@@ -1,5 +1,5 @@
 // LAF OS Library
-// Copyright (C) 2018-2021  Igara Studio S.A.
+// Copyright (C) 2018-2025  Igara Studio S.A.
 //
 // This file is released under the terms of the MIT license.
 // Read LICENSE.txt for more information.
@@ -11,6 +11,8 @@
 #include "os/color_space.h"
 
 #include <vector>
+
+#include <windows.h>
 
 namespace os {
